Separate Python & JS bindings
It turns out that Node can't make sense of the PyO3 bindings
when we build the NPM package with Neon, but it works fine when
I comment out all the Python decorator stuff. A full rewrite is
a bit much for a spike, but I at least separated the language
specific features to make it easier to switch between them.